requirements.txt
click==8.0.1
Flask==2.0.1
Flask-Scss==0.5
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
libsass==0.21.0
MarkupSafe==2.0.1
pymongo==3.12.0
pyScss==1.3.7
python-dotenv==0.19.0
six==1.16.0
Werkzeug==2.0.1
